+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Added `socket.disconnect()`
+
 ### Fixed
 - Creating an exponentially increasing retry interval caused a Lua error